eds_backend/requirements.txt
# ==================== CORE DJANGO ====================
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==25.1
drf-spectacular==0.28.0
asgiref==3.8.1
sqlparse==0.5.3

# ==================== DATABASE ====================
psycopg2-binary==2.9.9

# ==================== AUTHENTICATION ====================
django-allauth==65.4.0
PyJWT==2.10.1
bcrypt==5.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
pycparser==2.23

# ==================== EXCEL & DATA PROCESSING ====================
openpyxl==3.1.5
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.9.0.post0
et_xmlfile==2.0.0

# ==================== ENVIRONMENT & CONFIGURATION ====================
python-decouple==3.8
python-dotenv==1.0.0

# ==================== UTILITIES ====================
pytz==2025.2
tzdata==2025.1
requests==2.31.0
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Pillow==10.1.0

# ==================== API DOCUMENTATION ====================
inflection==0.5.1
uritemplate==4.2.0
PyYAML==6.0.2

# ==================== TESTING (Optional - can be removed in production) ====================
pytest==8.4.2
coverage==7.10.6
pluggy==1.6.0
iniconfig==2.1.0
packaging==25.0
